[LUCK-030] update wikit folder to include changes up to 13/08/2020
Issue:

We want to document team meetings, architecture used, changes made since the begining of the project and how the team collaborates together.

Solution:

Adding each file inside the wiki folder and explain in detail.

Risk:

In changedLog file, it doesn't include changes made for input file or meetings hold to implement it. Need Kevin or Tommy to add it in.
All links for Trello/google drive doc are shareable link, which means anyone with access to the wiki folder are able to view it. People with the link to Trello board will also have edit access.

Reviewed by:
Will, Kevin, Martin, Lucas, Tommy