[LUCK-012] change heuristic equation
Issue:

The original heuristic equation used - (endtime + weight/number of processor) doesn't necessary give a lower bound.
For example, consider task in processer 1: a ends at 2, and task b in processor 2: b ends at 5.
If we still have a third task c, based on the original equation, it will calculate the lower bound as (5 + weight of c / 2), however, it is possible for c to be scheduled directly after a in processor a.
The end time doesn't necessary need to be larger than 5.

Solution:

Change the heuristic equation to be (weight of all tasks + total idle time) / number of processors. This equation calculates the end time when there is perfect balanced loading on each processor.
After changing, this is able to produce the correct output for all example input files except Node_11_OutTree with 4 processors.

Risk:
This lower bound may not be good enough, so it cannot produce a solution for Node_11_OutTree with 4 processors.
Another heuristic equation which relates to critical path can be used as well to improve the lower bound.

Reviewed by:
Will, Lucas, Tommy, Kevin, Martin