[LUCK-012] Solve the problem that algorithm doesn't produce the correct optimal solution (#29)
Issue:
Two issues are identified in the algorithm:

1. 
When calculating the possible start time of a task node, _maxEndTime in SolutionNode may already have a value from the previous calculation.

2. 
The original heuristic equation used - (Endtime + weight/number of processors) doesn't necessarily give a lower bound.
For example, consider task "a" in processor 1 which ends at 2, and task "b" in processor 2 which ends at 5.
If we still have a third task "c", based on the current equation, it will calculate the lower bound as (5 + weight of c / 2), which is definitely larger than 5.
However, it is possible for c to be scheduled directly after "a" in processor 1. The end time doesn't necessarily need to be larger than 5.

Solution:

   - For the first issue identified:
    Reset _maxEndTime, _secondMaxEndTime and _latestProcessor every time the calculateStartTime function is called.

   - For heuristic equation:
Change the heuristic equation to be (weight of all tasks + total idle time) / number of processors. This equation calculates the end time when there is perfect load balance across all processors.
After changing, this is able to produce the correct output for all example input files except Node_11_OutTree with 4 processors, the program keeps running but cannot give an output.

   - To further improve the algorithm:
Add another equation - max(start time of each allocated task + its bottom level).
The final lower bound is the maximum between the result of two heuristic equations.
After adding this equation, the algorithm is able to produce the correct optimal solution for all example input files given.

To calculate the bottom level, each taskNode needs to maintain a list of outgoingEdges. Code related to creating outgoingEdges is copied from branch LUCK-012/Optimise-Algorithm and was originally written by Martin.

Risk:
/

Reviewed by:
Will, Lucas, Kevin, Tommy, Martin