[LUCK-012] refactor heuristic equation
Issue:

In the original implementation, every time we calculate the lower bound, we need to loop through all allocated tasks.
This significantly increases the runtime.

Solution:

Add another filed _criticalPath in the Processor which is the max(start time of an allocated task + bottom level of that task).
This field is updated every time a new task is added and it only stores the maxmimum value.
Therefore, when calculating the lower bound, we only need to loop through each processor, no longer need to loop through all allocated tasks.

Risk:
/

Reviewed by:
Will, Lucas, Martin, Tommy, Kevin