[LUCK-012] sort tasks in topological order
Sort tasks in topological order, this order is fixed in all unvisitedTasks list as well.
Therefore, when we checking if a task can be used to create partial solution, once we found a task that cannot be used to create solution due to unvisited parensts, there is no need to explore tasks after that.